+++ Source/Sitecore.FakeDb.Tests/Pipelines/GetTranslation/GetFakeTranslation.cs
@@ -1,0 +1,12 @@
+﻿namespace Sitecore.FakeDb.Tests.Pipelines.GetTranslation
+{
+  using Sitecore.Pipelines.GetTranslation;
+
+  public class GetFakeTranslation
+  {
+    public void Process(GetTranslationArgs args)
+    {
+      args.Result = args.Key + "*";
+    }
+  }
+}