+++ Source/Sitecore.FakeDb/Pipelines/GetTranslation/GetFakeTranslation.cs
@@ -1,0 +1,12 @@
+﻿namespace Sitecore.FakeDb.Pipelines.GetTranslation
+{
+  using Sitecore.Pipelines.GetTranslation;
+
+  public class GetFakeTranslation
+  {
+    public void Process(GetTranslationArgs args)
+    {
+      args.Result = args.Language.Name + ":" + args.Key;
+    }
+  }
+}